unkrautroboter_bilderkennung/pyproject.toml
[tool.poetry]
name = "unkrautroboter-bilderkennung"
version = "0.1.0"
description = "Bilderkennungssystem für einen Unkrautroboter"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
picamera2 = "*"
pillow = "*"
pyserial = "*"
ultralytics = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

